pan/mdg: Migrate liveness analysis to short masks
Cuts memory footprint in half.

+++ src/panfrost/midgard/compiler.h
@@ -186,7 +186,7 @@ typedef struct midgard_instruction {
 } midgard_instruction;
 
 /* Mask indexing into a 128-bit register */
-typedef uint16_t mir_mask;
+typedef uint8_t mir_mask;
 
 typedef struct midgard_block {
         pan_block base;
